temp

+++ src/playground_base/playground_base/base/topo/param/FBBoolParam.hpp
@@ -7,15 +7,15 @@ struct FBBoolParam
 {
   int ValueCount() const { return 2; }
 
-  float PlainToNormalized(bool plain) const;
-  bool NormalizedToPlain(float normalized) const;
+  double PlainToNormalized(bool plain) const;
+  bool NormalizedToPlain(double normalized) const;
 
   std::string PlainToText(bool plain) const;
   std::optional<bool> TextToPlain(std::string const& text) const;
 };
 
 inline bool
-FBBoolParam::NormalizedToPlain(float normalized) const
+FBBoolParam::NormalizedToPlain(double normalized) const
 {
   return normalized >= 0.5f;
 }
